aws setup - wip - it works!
