+++ configs/logging/log4j.default.properties
@@ -5,3 +5,5 @@ log4j.appender.file.File=${mist.home}/logs/mist.log
 log4j.appender.file.append=true
 log4j.appender.file.layout=org.apache.log4j.PatternLayout
 log4j.appender.file.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
+
+log4j.logger.net.schmizz.sshj=WARN
