+++ mist/master/src/main/scala/io/hydrosphere/mist/master/execution/ClustersService.scala
@@ -3,7 +3,7 @@ package io.hydrosphere.mist.master.execution
 import java.nio.file.Paths
 
 import akka.actor.ActorSystem
-import io.hydrosphere.mist.master.execution.aws.AWSEMRClusterRunner
+import io.hydrosphere.mist.master.execution.aws.EMRClusterRunner
 import io.hydrosphere.mist.master.{AWSEMRLaunchSettings, LauncherSettings}
 import io.hydrosphere.mist.master.models.{AWSEMRLaunchData, ContextConfig, ServerDefault}
 import io.hydrosphere.mist.utils.Logger
@@ -28,21 +28,19 @@ object ClustersService extends Logger{
     system: ActorSystem
   ): ClustersService = {
 
-    val runners = launchSettings.map({case (name, settings) => {
-      val runner = settings match {
-        case aws: AWSEMRLaunchSettings =>
-          AWSEMRClusterRunner.create(Paths.get(mistHome), spawn, aws, regHub, system)
-      }
-      name -> runner
-    }})
-
     new ClustersService {
+
+      val runners = launchSettings.map({case (name, settings) => {
+        val runner = settings match {
+          case aws: AWSEMRLaunchSettings => EMRClusterRunner.create(Paths.get(mistHome), spawn, aws, regHub, system)
+        }
+        name -> runner
+      }})
 
       override def start(id: String, ctx: ContextConfig): Future[Cluster] = {
         ctx.launchData match {
           case ServerDefault => serverDefault.run(id, ctx)
           case aws: AWSEMRLaunchData =>
-            logger.info(s"IN:${aws.launcherSettingsName} keys: ${runners.keys.toList}")
             runners.get(aws.launcherSettingsName) match {
               case Some(runner) => runner.run(id, ctx)
               case None => Future.failed(new RuntimeException(s"Unknown settings name ${aws.launcherSettingsName} for ctx ${ctx.name}"))
