+++ mist/master/src/main/scala/io/hydrosphere/mist/master/execution/workers/WorkerBridge.scala
@@ -2,7 +2,7 @@ package io.hydrosphere.mist.master.execution.workers
 
 import akka.actor.{Actor, ActorLogging, ActorRef, ActorRefFactory, Props, ReceiveTimeout, Terminated, Timers}
 import io.hydrosphere.mist.common.CommonData._
-import io.hydrosphere.mist.master.execution.WorkerLink
+import io.hydrosphere.mist.master.models.WorkerLink
 
 import scala.concurrent.{Future, Promise}
 import scala.concurrent.duration._
