+++ mist/master/src/test/scala/io/hydrosphere/mist/master/TestData.scala
@@ -3,7 +3,7 @@ package io.hydrosphere.mist.master
 import com.typesafe.config.ConfigFactory
 import io.hydrosphere.mist.common.CommonData.{Action, JobParams, RunJobRequest, WorkerInitInfo}
 import io.hydrosphere.mist.common.FunctionInfoData
-import io.hydrosphere.mist.master.execution.WorkerLink
+import io.hydrosphere.mist.master.models.WorkerLink
 import mist.api.ArgInfo
 import mist.api.data._
 import mist.api.encoding.defaultEncoders._
